Performing refactoring on the Restaurant Owners entity
