+++ src/main/java/br/com/techchallenge/domain/entity/DonoRestaurante.java
@@ -1,6 +1,5 @@
 package br.com.techchallenge.domain.entity;
 
-import jakarta.persistence.Column;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -14,7 +13,7 @@ public class DonoRestaurante {
 
     private Long id;
     private String nome;
-    private String endereco;
+    private Endereco endereco;
     private String email;
     private String login;
     private String senha;
