+++ src/main/java/br/com/techchallenge/domain/entity/Restaurante.java
@@ -1,6 +1,5 @@
 package br.com.techchallenge.domain.entity;
 
-import br.com.techchallenge.domain.useCase.donoRestaurante.DonoRestaurante;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
