+++ src/main/java/br/com/techchallenge/domain/gateway/donoRestaurante/BuscarTodosDonoRestauranteInterface.java
@@ -1,6 +1,6 @@
 package br.com.techchallenge.domain.gateway.donoRestaurante;
 
-import br.com.techchallenge.domain.useCase.donoRestaurante.DonoRestaurante;
+import br.com.techchallenge.domain.entity.DonoRestaurante;
 
 import java.util.List;
