+++ src/main/java/br/com/techchallenge/domain/useCase/restaurante/CadastrarRestauranteUseCase.java
@@ -1,10 +1,10 @@
 package br.com.techchallenge.domain.useCase.restaurante;
 
+import br.com.techchallenge.domain.entity.DonoRestaurante;
 import br.com.techchallenge.domain.entity.Restaurante;
 import br.com.techchallenge.domain.exception.DonoRestauranteNaoExisteException;
 import br.com.techchallenge.domain.exception.RestauranteJaCadastradoException;
 import br.com.techchallenge.domain.gateway.restaurante.RestauranteSalvarInterface;
-import br.com.techchallenge.domain.useCase.donoRestaurante.DonoRestaurante;
 
 import java.util.Optional;
